methodjit: introduce typed ExitKind for the JIT exit ABI
Replace the untyped int exit-code constants (ExitNormal=0 ...
ExitCoroutineYieldFast=10) with a typed `ExitKind int64` enum plus a
centralized table in exit_kind.go (allExitKinds, ExitKind.String(),
IsKnownExitKind). ExecContext.ExitCode, ExecContext.NativeCalleeExitCode,
NativeCallExitFrame.NativeCalleeExitCode, exitStatsKey.code and the
exit-resume-check key are now ExitKind so the type checker catches invalid
exit-kind values and missing dispatch handlers.

Behavior-preserving and ABI-unchanged: ExitKind is int64-backed, so the
ExecContext.ExitCode field keeps the same 8-byte size, alignment and offset,
and the generated ARM64 STR/LDR/LoadImm64/CMPimm of the exit code emit the
same instructions and numeric literals (explicit int64()/uint16() casts added
only at the asm and JSON boundaries).

Adds exit_kind_test.go: pins the numeric ABI values + field size/offset, and an
exhaustiveness check that every enumerated kind has a Tier 1 / Tier 2 / native
dispatch handler.

+++ internal/methodjit/emit_arith.go
@@ -622,7 +622,7 @@ func (ec *emitContext) emitIntModZeroDeopt() {
 	asm := ec.asm
 	asm.LoadImm64(jit.X0, -1)
 	asm.STR(jit.X0, mRegCtx, execCtxOffDeoptInstrID)
-	asm.LoadImm64(jit.X0, ExitDeopt)
+	asm.LoadImm64(jit.X0, int64(ExitDeopt))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
@@ -715,7 +715,7 @@ func (ec *emitContext) emitInt48OverflowCheck(result jit.Reg, instr *Instr) {
 		asm.LoadImm64(jit.X0, int64(instr.ID))
 		asm.STR(jit.X0, mRegCtx, execCtxOffDeoptInstrID)
 	}
-	asm.LoadImm64(jit.X0, ExitDeopt)
+	asm.LoadImm64(jit.X0, int64(ExitDeopt))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
