+++ internal/methodjit/emit_deopt.go
@@ -25,7 +25,7 @@ func (ec *emitContext) emitDeopt(instr *Instr) {
 		asm.LoadImm64(jit.X0, int64(instr.ID))
 		asm.STR(jit.X0, mRegCtx, execCtxOffDeoptInstrID)
 	}
-	asm.LoadImm64(jit.X0, ExitDeopt)
+	asm.LoadImm64(jit.X0, int64(ExitDeopt))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
