+++ internal/methodjit/emit_table_ctor.go
@@ -166,7 +166,7 @@ func (ec *emitContext) emitNewFixedTable2Exit(instr *Instr, resultSlot int) {
 	asm.STR(jit.X0, mRegCtx, execCtxOffTableExitID)
 
 	ec.emitSetResumeNumericPass()
-	asm.LoadImm64(jit.X0, ExitTableExit)
+	asm.LoadImm64(jit.X0, int64(ExitTableExit))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
@@ -377,7 +377,7 @@ func (ec *emitContext) emitNewFixedTableNExit(instr *Instr, resultSlot int) {
 	asm.STR(jit.X0, mRegCtx, execCtxOffTableExitID)
 
 	ec.emitSetResumeNumericPass()
-	asm.LoadImm64(jit.X0, ExitTableExit)
+	asm.LoadImm64(jit.X0, int64(ExitTableExit))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
