+++ internal/methodjit/emit_table_field_load.go
@@ -833,7 +833,7 @@ func (ec *emitContext) emitGetFieldExit(instr *Instr) {
 
 	// Set ExitCode = ExitTableExit and return to Go.
 	ec.emitSetResumeNumericPass()
-	asm.LoadImm64(jit.X0, ExitTableExit)
+	asm.LoadImm64(jit.X0, int64(ExitTableExit))
 	asm.STR(jit.X0, mRegCtx, execCtxOffExitCode)
 	if ec.numericMode {
 		asm.B("num_deopt_epilogue")
